Eliminating unnecessary header
The <h1> was a bit redundant before, because Tumblr puts the title of the page as an <h3> above it, but without it, the div would vanish.  I figured out the div needed to be wrapped in a paragraph for Tumblr's sake, and now we don't need the <h1> anymore.