adding option to treat all tags as lowercase
Tumblr itself seems to treat tags in a case-insensitive way, and the convention of tumblr users seems to tend toward writing tags in all lower-case.  If caseInsensitive is set to true, all tags will be counted and displayed as if they are all lower-case (I think this would be the generally preferred usage).  Otherwise, different capitalizations will be counted and displayed in the list separately (might be useful for seeing different ways a tag has been typed).